+++ src/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @turtledev/api
 
+## 1.4.1
+
+### Patch Changes
+
+- Fixes to themes & route
+
 ## 1.4.0
 
 ### Minor Changes